+++ src/runtime-rs/crates/hypervisor/src/qemu/mod.rs
@@ -5,7 +5,6 @@
 
 mod cmdline_generator;
 mod inner;
-mod network;
 
 use crate::device::DeviceType;
 use crate::hypervisor_persist::HypervisorState;
